feat: moved in to `generator` package
