+++ src/main/java/vstu/isd/notebin/generator/UrlGeneratorImpl.java
@@ -1,11 +1,10 @@
-package vstu.isd.notebin.util;
+package vstu.isd.notebin.generator;
 import org.springframework.stereotype.Component;
 
 import java.util.UUID;
 
 @Component
 public class UrlGeneratorImpl implements UrlGenerator {
-
     @Override
     public String generateUrl() {
         return UUID.randomUUID().toString();
